Checkstyle fix
