+++ src/main/java/core/basesyntax/Application.java
@@ -1,0 +1,14 @@
+package core.basesyntax;
+
+import core.basesyntax.controller.ConsoleHandler;
+import core.basesyntax.lib.Injector;
+import java.lang.reflect.InvocationTargetException;
+
+public class Application {
+    public static void main(String[] args) throws InvocationTargetException,
+            NoSuchMethodException, InstantiationException, IllegalAccessException {
+        ConsoleHandler handler = (ConsoleHandler) Injector
+                .getInstance(ConsoleHandler.class);
+        handler.handle();
+    }
+}
