+++ src/main/java/core/basesyntax/dao/UserDaoImpl.java
@@ -1,9 +1,8 @@
-package core.basesyntax.controller.dao;
+package core.basesyntax.dao;
 
-import core.basesyntax.controller.db.StorageUsers;
-import core.basesyntax.controller.model.User;
+import core.basesyntax.db.StorageUsers;
 import core.basesyntax.lib.Dao;
-
+import core.basesyntax.model.User;
 import java.util.List;
 
 @Dao
