+++ src/main/java/core/basesyntax/db/StorageUsers.java
@@ -1,6 +1,6 @@
-package core.basesyntax.controller.db;
+package core.basesyntax.db;
 
-import core.basesyntax.controller.model.User;
+import core.basesyntax.model.User;
 import java.util.ArrayList;
 import java.util.List;
