+++ src/main/java/core/basesyntax/exception/NotHaveAnnotation.java
@@ -1,0 +1,7 @@
+package core.basesyntax.exception;
+
+public class NotHaveAnnotation extends RuntimeException {
+    public NotHaveAnnotation(String message) {
+        super(message);
+    }
+}
