+++ src/main/java/core/basesyntax/lib/Inject.java
