--- src/main/java/exception/NotHaveAnnotation.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class NotHaveAnnotation extends RuntimeException{
-    public NotHaveAnnotation(String message) {
-        super(message);
-    }
-}
